Split out schema statics
